issue #100 refactor: post change to story, directory
