+++ src/components/Landing/Main/Main.style.ts
