+++ src/components/Landing/Main/Main.tsx
@@ -1,6 +1,8 @@
 import Flex from "@/components/common/Flex/Flex";
-import { Searchbar, Sidebar, Post } from "@/components/Landing";
-import { mainStyle } from "@/components/Landing/Main/Main.styles";
+import { Searchbar, Sidebar } from "@/components/Landing";
+import { Story } from "@/components/Story";
+
+import { mainStyle } from "@/components/Landing/Main/Main.style";
 
 const Main = () => {
 	return (
@@ -8,7 +10,7 @@ const Main = () => {
 			<Flex styles={{ gap: "30px" }}>
 				<section>
 					<Searchbar defaultText="전체검색" />
-					<Post />
+					<Story />
 				</section>
 				<Sidebar />
 			</Flex>
