+++ src/components/Landing/index.tsx
@@ -1,10 +1,7 @@
 import Bio from "@/components/Landing/Bio/Bio";
 import Main from "@/components/Landing/Main/Main";
-import Post from "@/components/Landing/Post/Post";
-import PostCard from "@/components/Landing/Post/PostCard";
-import PostTitle from "@/components/Landing/Post/PostTitle";
 import Searchbar from "@/components/Landing/Searchbar/Searchbar";
 import Sidebar from "@/components/Landing/Sidebar/Sidebar";
 import SortButton from "@/components/Landing/SortButton/SortButton";
 
-export { Bio, Main, Post, PostCard, PostTitle, Searchbar, Sidebar, SortButton };
+export { Bio, Main, Searchbar, Sidebar, SortButton };
