+++ src/components/Planning/Calendar/CalendarCard/CalendarCard.tsx
@@ -1,11 +1,10 @@
 import type { PropsWithChildren } from "react";
 
-import { format, isSameDay } from "date-fns";
-
 import { Box, Flex, Text } from "@/components/common";
 import MoreButton from "@/components/Planning/Calendar/CalendarCard/MoreModal/MoreButton";
 import MoreModal from "@/components/Planning/Calendar/CalendarCard/MoreModal/MoreModal";
 import ScheduleModal from "@/components/Planning/Calendar/CalendarCard/ScheduleModal/ScheduleModal";
+import { format, isSameDay } from "date-fns";
 
 import { MAX_CALENDAR_CONTENT } from "@/constants/calendar";
