+++ src/components/Planning/Calendar/CalendarCard/MoreModal/MoreModal.tsx
@@ -1,9 +1,8 @@
 import { useRef } from "react";
 
-import { format } from "date-fns";
-
 import { Box, Flex, Text } from "@/components/common";
 import ScheduleModal from "@/components/Planning/Calendar/CalendarCard/ScheduleModal/ScheduleModal";
+import { format } from "date-fns";
 
 import { MAX_CALENDAR_CONTENT } from "@/constants/calendar";
