+++ src/components/Planning/Calendar/CalendarHeader/CalendarHeader.tsx
@@ -1,9 +1,8 @@
-import { format } from "date-fns";
-
 import LeftArrowIcon from "@/assets/svg/left-arrow.svg?react";
 import RightArrowIcon from "@/assets/svg/right-arrow.svg?react";
 
 import { Flex, Heading } from "@/components/common";
+import { format } from "date-fns";
 
 import {
 	flexStyle,
