+++ src/components/Story/Story.style.ts
