+++ src/components/Story/StoryDetail/Profile.style.ts
