+++ src/components/Story/StoryDetail/StoryDetail.style.ts
