+++ src/components/Story/index.tsx
@@ -1,0 +1,5 @@
+import Story from "@/components/Story/Story";
+import StoryCard from "@/components/Story/StoryCard";
+import StoryTitle from "@/components/Story/StoryTitle";
+
+export { Story, StoryCard, StoryTitle };
